Add Pydantic v2 data model
This update adds a Pydantic v2 data model in the datamodel module along
with unit tests for the new functionality. A new CLI command "validate"
supports loading a YAML file in the top-level Schema class and
validating it. Validation errors will result in a non-zero return code.
Any number of YAML files may be provided at the command line and
wildcards may be used to test all the files in a directory.

The data model explicitly defines fields for all allowed keys, such
that the usage of undefined keys, including misspellings or unknown
annotations, will result in validation errors. All optional annotations
currently found in the SDM Schemas YAML files are supported by the data
model. Units and IVOA UCDs are checked using modules from astropy.
Validation of the contents of other fields is initially loose, some
accepting any string value. It is forseen that additional validation
rules for fields and data model classes will be added in the future.