Fix referencing problem with foreign key objects
This is a fix for an issue where the "load-tap" command was failing due
to how jsonld was configured. The default when calling jsonld.frame()
is embedding the object once. The visitor was processing the first
constraint under the assumption that it would see a column object. But
for subsequent calls, it was only seeing the reference string value.
The method is now configured to use the "@always" setting when loading
data into TAP so that the object itself will always be seen. This was
changed only for this particular command by adding a switch to the
cli._normalize() method which maintains the jsonld default.

The tap module is changed based on the assumption that the column
object values will be seen instead of references. The "@id" value from
the column is used as a lookup into the dictionary which contains the
TAP column representation.