Modify TAP visitor so constraints do not depend on column order
The TapLoadingVisitor was modified so that constraints are created in a
second pass over the schema. This ensures that columns which may be
referenced have been defined. The change was tested successfully using
"load-tap" from the command line, with and without the "--dry-run"
argument. The TAP data loaded into a MySQL database correctly when
using the non-mock connection.