Check for duplicates when validating schema
The Pydantic model was missing a check for duplicate IDs, which can
lead to various errors when using other Felis modules such as tap. This
will make a list of duplicates as they are found and then raise an
error. The entire list will then be printed out after validation.