Add check for duplicate IDs when validating a schema
The function for creating the ID map was not working and was rewritten
to use a visitor pattern instead of recursion. Duplicates are added to
a set as the schema is visited. The model_validator will raise an error
if duplicates are found and print a complete list of the duplicated
IDs.