Add metadata module and implementation of SchemaMetaData class
Add a module with a new class that encapsulates the creation of
sqlalchemy MetaData from a Pydantic Schema object, along with a test
that dumps the corresponding SQL DDL statements to a file. The logic
for translating into SQA was primarily derived from the existing
visitor class in sql.py, which this is intended to replace eventually.