Add flags for applying the schema name to the metadata and tables
In some cases, it is not desirable to apply the schema name to either
the metadata or the tables. Add two flags to the builder for
controlling this behavior. The flags are true by default and can be
set to false if the user does not want the schema to be applied.