Refactor tap module to use the Pydantic data model
The tap module is updated to use the Pydantic data model, instead of
the raw YAML. The "load-tap" command in the command line interface
was also updated to use the Pydantic schema object. The PYLD
transformations were removed from the cli function, as they seemed to
be unnecessary. A few changes were made to the Pydantic data model to
support setting of reasonable defaults. A "votable_datatype" annotation
was added to Colum. The Column datatype was changed to return the enum
instead of a string, as this setting seemed to confuse mypy. Tests were
changed to conform to the new tap module and the minor changes to the
data model. Testing showed that the schema was loaded correctly into
a live PostgreSQL database.