Improve support of TEXT type (DM-43097)
Felis insists that `text` datatype is sized and required `length` to be
specified. But it maps `text` to MySQL `LONGTEXT` and PostgreSQL `TEXT`
types which do not accept length specification. This patch adds a
workaround for those two types which drops length argument. A unit test
was added as well to trigger various cases for `text` and other sized types

`tests/data/sales.yaml` specified mysql datatype override as `INT`
which does not exist (in sqlalchemy at least), it needs to be `INTEGER`.