Use Pydantic context for validation flags rather than global variables
When the validation flags are represented with globals, there can be
issues with global state not being reset correctly, especially when
running tests.

Instead of having these settings as class variables on the Schema,
make them instance variables and pass them down to the appropriate
objects during initialization. This ensures that no unwanted state is
persisted between model validations.