Handle default values properly when defining columns in metadata
The 'value' field in the schema can be used to provide a default column
value or `server_default` in a SQLAlchemy column. This needs to be
handled properly based on several different cases. String values need
to be quoted unless they represent SQL functions, in which case they
are unquoted. Other values such as numerics are passed without quotes.
These cases were checked using some new default values that were added
to a test schema.