Add proper handling and validation of column value field
This adds validation of the column `value` field, which provides a
default that is passed to `server_default` when creating a SQLAlchemy
column.

The `MetaDataBuilder` was updated to properly handle these values
depending on their type and content. Strings are quoted, except when
they represent SQL functions such as `CURRENT_TIMESTAMP`.

Tests were added to check that the validation functions properly for
all of the Felis datatypes.