Add validation of length for sized datatypes
Add a standalone validation function that checks if a sized type has
a length and raises an error if not.

Add a requirement that the length field in the data model is greater
than zero.

Add lengths to some string columns in the tests, as length is
always required now for `Column` objects.