Add new tap_schema module
Add a new tap_schema module which performs the same functions as the
existing tap module and is designed to replace it.

The new module has an object-oriented organization rather than a mix
of functions and classes as in the old module. The new control flow
and translation of the data from the schema should be easier to follow
now than before, as well as more convenient to use via the Python API.

The two main classes are the TableManager which creates the SQLAlchemy
metadata for the TAP_SCHEMA database and the DataLoader, which loads
the schema data.

New test modules were added, one of which uses testing.postgresql to
create a PostgreSQL database, allowing the classes to be easily tested
against a more production-like database than SQLite.