Fix arraysize not being set for text columns
Rewrite the logic for setting arraysize so that it is simpler and does
not check if the Felis type is sized.

Instead set arraysize from the column or use it's length if arraysize
is not set. For timestamp or text columns, set the arraysize to '*' if
it was not set explicitly in the schema.