Add Schema validator that automatically generates object IDs
Implement a pre-validator which creates object IDs where they are
missing on Schema objects.

The values of the generated IDs for schema, table and column objects
are constructed based on the common usage patterns from SDM Schemas.

Constraints and indexes use the name field as the ID.

The `--id-generation` flag must be used from the command-line to enable
this feature, as it is off by default.