Cleanup unit tests of constraints
Remove duplicate checks on name and id values.

Normalize the object names and ids to snakecase.

Add a function to test the base Constraint object, which is never
instantiated, but has deferrable and initially that need to be checked.

Move checks on name and id to the test case for the base class.

Fix some erroneous variable names.